Update of the parser
Updating the parser to avoid some issues with the LABEL position in the Tree.
