+++ Parsing/parsing.h
@@ -1,55 +1,86 @@
-#ifndef PARSING_H
-#define PARSING_H
-#include "token.h"
+typedef struct astNode {
+    token_t token;
+    struct astNode **children;
+    int numChildren;
+} astNode_t;
 
-
-typedef struct node
-{
-    token_t tokens;
-    struct node *left;
-    struct node *right;
-}node_t;
-
-node_t* newNode(token_t token) {
-    node_t* node = (node_t*)malloc(sizeof(node_t));
-    node->tokens = token;
-    node->left = NULL;
-    node->right = NULL;
+astNode_t *createNode(token_t token) {
+    astNode_t *node = malloc(sizeof(astNode_t));
+    node->token = token;
+    node->children = NULL;
+    node->numChildren = 0;
     return node;
 }
 
-node_t* createTree(token_t* tokens, int numTokens) {
-    if (numTokens == 0) {
-        return NULL;
+void addChild(astNode_t *parent, astNode_t *child) {
+    parent->children = realloc(parent->children, sizeof(astNode_t *) * (parent->numChildren + 1));
+    parent->children[parent->numChildren] = child;
+    parent->numChildren++;
+}
+
+astNode_t *buildAST(token_t *tokens, int numTokens) {
+    astNode_t *root = createNode((token_t){"ROOT", "ROOT", 0, 0});
+    astNode_t *currentParent = root;
+    astNode_t **parentStack = malloc(sizeof(astNode_t *) * numTokens);
+    int stackTop = 0;
+    parentStack[stackTop] = root;
+
+    bool inCall = false;
+    astNode_t *lastCallNode = NULL;
+    int lastInstructionRow = -1;
+
+    for (int i = 0; i < numTokens; i++) {
+        if (strcmp(tokens[i].type, "VOID") == 0) continue;
+        astNode_t *currentNode = createNode(tokens[i]);
+
+        if (strcmp(tokens[i].type, "LABEL") == 0 && inCall) {
+            addChild(lastCallNode, currentNode);
+            inCall = false;
+        } else if (strcmp(tokens[i].type, "LABEL") == 0) {
+            addChild(currentParent, currentNode);
+            parentStack[++stackTop] = currentNode;
+            currentParent = currentNode;
+        } else if (strcmp(tokens[i].value, "RET") == 0 || strcmp(tokens[i].value, "END") == 0) {
+            addChild(currentParent, currentNode);
+            currentParent = parentStack[--stackTop];
+        } else {
+            if (tokens[i].column == 1 && tokens[i].row != lastInstructionRow) {
+                currentParent = parentStack[stackTop];
+            }
+            addChild(currentParent, currentNode);
+            if (strcmp(tokens[i].type, "INSTRUCTION") == 0) {
+                lastInstructionRow = tokens[i].row;
+                if (strcmp(tokens[i].value, "CALL") == 0) {
+                    inCall = true;
+                    lastCallNode = currentNode;
+                } else {
+                    currentParent = currentNode;
+                }
+            }
+        }
     }
-
-    node_t* root = newNode(tokens[0]);
-
-    // Create left and right subtrees from the rest of the tokens
-    root->left = createTree(tokens + 1, numTokens / 2);
-    root->right = createTree(tokens + numTokens / 2 + 1, numTokens - numTokens / 2 - 1);
-
+    free(parentStack);
     return root;
 }
 
-void printNode(node_t* node) {
-    if (node == NULL) {
-        return;
-    }
+void printAST(astNode_t *node, int depth) {
+    if (node == NULL) return;
 
-    printf("Token: %s \t %s \t %d \t %d \n", node->tokens.type, node->tokens.value, node->tokens.row, node->tokens.column);
-    
-    printNode(node->left);
-    
-    printNode(node->right);
+    for (int i = 0; i < depth; i++) printf("  ");
+    printf("Token(\"%s\", \"%s\", %d, %d)\n", node->token.type, node->token.value, node->token.row, node->token.column);
+
+    for (int i = 0; i < node->numChildren; i++) {
+        printAST(node->children[i], depth + 1);
+    }
 }
 
+void freeAST(astNode_t *node) {
+    if (node == NULL) return;
 
+    for (int i = 0; i < node->numChildren; i++) {
+        freeAST(node->children[i]);
+    }
 
-#endif
-
-
-
-
-
-
+    free(node->children);
+    free(node);
+}
